Refactor backup settings - move shared components to common-ui
