+++ src/common-ui/components/loading-blocker.jsx
