+++ src/options/backup/container.tsx
@@ -4,11 +4,11 @@ import { remoteFunction } from 'src/util/webextensionRPC'
 import analytics from 'src/analytics'
 import { default as Overview } from './screens/overview'
 import { default as RunningBackup } from './screens/running-backup'
-import { default as OnboardingWhere } from './screens/onboarding-1-where'
-import OnboardingHow from './screens/onboarding-2-how'
-import { default as OnboardingSize } from './screens/onboarding-3-size'
+import { default as OnboardingWhere } from './screens/setup-location'
+import OnboardingHow from './screens/setup-manual-or-automatic'
+import { default as OnboardingSize } from './screens/setup-size'
 import { BackupHeader } from './components/backup-header'
-import LoadingBlocker from './components/loading-blocker'
+import LoadingBlocker from '../../common-ui/components/loading-blocker'
 import * as logic from 'src/options/backup/container.logic'
 import RestoreWhere from 'src/options/backup/screens/restore-where'
 import RestoreRunning from 'src/options/backup/screens/restore-running'
