+++ src/options/backup/screens/overview.tsx
@@ -2,8 +2,8 @@ import moment from 'moment'
 import React from 'react'
 import classNames from 'classnames'
 import { remoteFunction } from 'src/util/webextensionRPC'
-import SmallButton from '../components/small-button'
-import LoadingBlocker from '../components/loading-blocker'
+import SmallButton from '../../../common-ui/components/small-button'
+import LoadingBlocker from '../../../common-ui/components/loading-blocker'
 import RestoreConfirmation from '../components/restore-confirmation'
 import { browser } from 'webextension-polyfill-ts'
 import { SubscribeModal } from 'src/authentication/components/Subscription/SubscribeModal'
