+++ src/options/backup/screens/setup-size.css
